conversai-backend/requirements.txt
Flask==2.2.2
flask-cors==3.0.10
cohere==4.4.0
python-dotenv==0.21.0
